refactor: move task and experiment utilities to the right place
